+++ R/magma_executable.R
@@ -14,8 +14,11 @@ magma_executable <- function(exec_dir = find_install_dir(verbose=FALSE),
     #### Find the executable file once it's been downloaded ####
     magma_x <- list.files(exec_dir,"^magma$", 
                           recursive = TRUE, full.names = TRUE) 
+    magma_x <- magma_x[!is.na(magma_x)]
+    
     if(!is.null(version)){
         v_search <- grep(version,magma_x, value = TRUE, ignore.case = TRUE)
+        v_search <- v_search[!is.na(v_search)]
         if(length(v_search)>0) {
             magma_x <- v_search
         } else {
@@ -33,5 +36,5 @@ magma_executable <- function(exec_dir = find_install_dir(verbose=FALSE),
         stop_msg <- paste("Cannot find MAGMA executable.")
         stop(stop_msg)
     }
-    return(magma_x)
+    return(normalizePath(magma_x))
 }